create a new temperature point and station if new
When we submit a temperature point we create a new station if the
station provided doesn't exist. This is mostly just to ease
testing/validation. In an ideal scenario we'd have a separate set of
endpoints to create stations + generate auth tokens which would then be
required during all other operations.